requirements.txt
fastapi==0.115.14
uvicorn==0.22.0
gradio==5.35.0
langchain==0.0.299
sentence-transformers==2.2.2
transformers==4.33.3
huggingface-hub>=0.30.0,<1.0.0
faiss-cpu==1.7.3
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.11
newspaper3k==0.2.8
PyPDF2==3.0.0
Pillow==10.0.1
APScheduler==3.11.0
schedule==1.2.2
numpy==1.25.2
matplotlib==3.8.2
python-telegram-bot==20.5
discord.py==2.5.1